+++ src/basic/__tests__/basic.test.tsx
@@ -10,7 +10,7 @@ import {
 } from '@testing-library/react';
 import { CartPage } from '../../refactoring/pages/CartPage';
 import { AdminPage } from '../../refactoring/pages/AdminPage';
-import { CartItem, Coupon, Product } from '../../types';
+import { CartItem, Coupon, Product } from '../../refactoring/types';
 import { useCart, useCoupons, useProducts } from '../../refactoring/hooks';
 import * as cartUtils from '../../refactoring/models/cart';
