+++ src/origin/components/AdminPage.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { Coupon, Discount, Product } from '../../types';
+import { Coupon, Discount, Product } from '../../refactoring/types';
 
 interface Props {
   products: Product[];
