+++ src/refactoring/data/products.ts
@@ -1,4 +1,4 @@
-import { Product } from '../../types';
+import { Product } from '../types';
 
 export const Products: Product[] = [
   {
