+++ src/refactoring/hooks/useCoupon.ts
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { Coupon } from '../../types.ts';
+import { Coupon } from '../types';
 
 export const useCoupons = (initialCoupons: Coupon[]) => {
   const [coupons, setCoupons] = useState<Coupon[]>(initialCoupons);
