+++ src/refactoring/models/cart.ts
@@ -1,4 +1,4 @@
-import { CartItem, Coupon } from '../../types';
+import { CartItem, Coupon } from '../types';
 
 export const calculateItemTotal = (item: CartItem) => {
   const { quantity, product } = item;
