+++ src/refactoring/utils/discount.ts
@@ -1,4 +1,4 @@
-import { CartItem } from '../../types.ts';
+import { CartItem } from '../types';
 
 // 주어진 장바구니 아이템에 대해 적용된 할인율을 계산하는 함수입니다.
 export const getAppliedDiscount = (item: CartItem) => {
